requirements.txt
Flask==3.0.3
boto3==1.18.0
psycopg2-binary==2.9.3
aws-wsgi


